OCR_Py/requirements.txt
pytesseract==0.3.8
pandas==1.1.3
parsel==1.5.2
numpy==1.19.2
opencv_python==4.5.2.54
easyocr==1.3.2
scipy==1.5.2
RandomWords==0.3.0
Pillow==8.3.1
